Refactor imports to use updated User entity path in mappers, services, and controllers
