+++ Backend/src/main/java/com/pg/mbti/mappers/UserMapper.java
@@ -1,7 +1,7 @@
 package com.pg.mbti.mappers;
 
 import com.pg.mbti.dto.UserProfileDto;
-import com.pg.mbti.entity.user.User;
+import com.pg.mbti.entity.User;
 import org.springframework.stereotype.Component;
 
 @Component
