+++ Backend/src/main/java/com/pg/mbti/repositories/UsersRepository.java
@@ -1,6 +1,6 @@
 package com.pg.mbti.repositories;
 
-import com.pg.mbti.entity.user.User;
+import com.pg.mbti.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
